requirements.txt
aiofiles==0.5.0
aiosqlite==0.13.0
alembic==1.4.2
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
Cerberus==1.3.2
click==7.1.2
databases==0.3.2
fake-rpi==0.7.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
httptools==0.1.1
hyperframe==5.2.0
Mako==1.1.2
MarkupSafe==1.1.1
numpy==1.18.4
pathspec==0.8.0
priority==1.3.0
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.3.1
regex==2020.5.14
six==1.14.0
SQLAlchemy==1.3.17
starlette==0.13.4
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
uvicorn==0.11.5
uvloop==0.14.0
websockets==8.1
wsproto==0.15.0
